pyproject.toml
[tool.poetry]
name = "homework8"
version = "0.1.0"
description = "MongoDB, RabbitMQ"
package-mode = false
authors = ["Yevhenii Baranov <j52b10i78@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mongoengine = "^0.28.2"
python-dotenv = "^1.0.1"
pika = "^1.3.2"
faker = "^26.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
